# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CommentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasePageResultCommentDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"limit": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/CommentDto"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"subject": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"subject",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasePageResultPostDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"limit": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/PostDto"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPost": {
				"properties": {
					"name": {
						"type": "string"
					},
					"subject": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"subject",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"firstName",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Registration": {
				"properties": {
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Login": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasePageResultUserDto": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"limit": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/UserDto"
						},
						"type": "array"
					}
				},
				"required": [
					"count",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUser": {
				"properties": {
					"firstName": {
						"type": "string",
						"nullable": true
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/comments": {
			"get": {
				"operationId": "GetAllComments",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasePageResultCommentDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "my",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "postId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Comment"
							}
						}
					}
				}
			}
		},
		"/api/comments/{id}": {
			"get": {
				"operationId": "GetCommentById",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateComment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Comment"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteComment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/posts": {
			"get": {
				"operationId": "GetAllPosts",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasePageResultPostDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Posts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "my",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Posts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPost"
							}
						}
					}
				}
			}
		},
		"/api/posts/{id}": {
			"get": {
				"operationId": "GetPostById",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Posts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPost"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Posts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/auth/registration": {
			"post": {
				"operationId": "Registration",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Registration"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login"
							}
						}
					}
				}
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasePageResultUserDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"role": {
													"type": "string"
												}
											},
											"required": [
												"role"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/UserDto"
										}
									]
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUser"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:5000/"
		}
	]
}